requirements.txt
# Install the latest version of the package
requests

# Install Python environment variables reader
python-dotenv>=0.19.2